requirements.txt
# Flask and Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
gunicorn==21.2.0

# Machine Learning & NLP
scikit-learn==1.3.2
numpy==1.26.2

# Web Scraping
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3

# Data Processing
python-dotenv==1.0.0

# Utilities
joblib==1.3.2
urllib3==2.1.0

# Production Server
gunicorn==21.2.0

# Testing
pytest==7.4.3
